Move protogalaxy
