+++ barretenberg/cpp/src/barretenberg/protogalaxy/combiner.test.cpp
@@ -1,6 +1,6 @@
 #include "barretenberg/flavor/ultra.hpp"
-#include "barretenberg/honk/proof_system/protogalaxy_prover.hpp"
 #include "barretenberg/honk/utils/testing.hpp"
+#include "barretenberg/protogalaxy/protogalaxy_prover.hpp"
 #include "barretenberg/relations/relation_parameters.hpp"
 #include "barretenberg/sumcheck/instance/instances.hpp"
 #include <gtest/gtest.h>
