+++ barretenberg/cpp/src/barretenberg/protogalaxy/folding_result.hpp
