+++ barretenberg/cpp/src/barretenberg/protogalaxy/protogalaxy_prover.cpp
