+++ barretenberg/cpp/src/barretenberg/protogalaxy/protogalaxy_verifier.cpp
